+++ Realm/Tests/ObjectTests.m
@@ -1154,8 +1154,9 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
     XCTAssertEqual([objects count], 2U, @"Should have 2 objects");
     XCTAssertEqual([(PrimaryStringObject *)objects[0] intCol], 3, @"Value should be 3");
 
-    // upsert on non-primary key object shoudld throw
+    // upsert on non-primary key object should throw
     XCTAssertThrows([StringObject createOrUpdateInDefaultRealmWithObject:@[@"string"]]);
+    XCTAssertThrows([StringObject createOrUpdateInRealm:realm withObject:@[@"string"]]);
 
     [realm commitWriteTransaction];
 }