+++ CHANGELOG.md
@@ -7,13 +7,18 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 
-* None.
+* Calling `createInDefaultRealmWithObject:`, `createInRealm:withObject:`,
+  `createOrUpdateInDefaultRealmWithObject:` or `createOrUpdateInRealm:withObject:`
+  is a no-op if the argument is an RLMObject of the same type as the receiver
+  and is already backed by the target realm.
 
 ### Bugfixes
 
 * Fix incorrect column type assertions when the first Realm file opened is a
   read-only file that is missing tables.
 * Throw an exception when adding an invalidated or deleted object as a link.
+* Throw an exception when calling `createOrUpdateInRealm:withObject:` when the
+  receiver has no primary key defined.
 
 0.90.1 Release notes (2015-01-22)
 =============================================================